Delegate course and courseOffering look-ups to CourseRepository and move implementations to InMemoryCourseRepository.
