+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/domain/CourseRepository.kt
@@ -1,3 +1,10 @@
 package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.domain
 
-interface CourseRepository
+import java.util.*
+
+interface CourseRepository {
+  fun allCourses(): List<CourseEntity>
+  fun course(courseId: UUID): CourseEntity?
+  fun offeringsForCourse(courseId: UUID): List<Offering>
+  fun courseOffering(courseId: UUID, offeringId: UUID): Offering?
+}
